Restructuring of RISC-V code and various bugfixes
